services/ai-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
torch==2.1.2
torchvision==0.16.2
transformers==4.36.2
Pillow==10.1.0
numpy==1.26.2
